+++ src/dm/factory.js
@@ -1,21 +1,30 @@
 var inherits = require("inherits-js"),
+    _        = require("./utils"),
     Factory;
 
-Factory = function() {
-    //
+/**
+ * @abstract
+ * @class Factory
+ * @constructor
+ */
+Factory = function(options) {
+    this.options = _.extend({}, this.constructor.DEFAULTS, options || {});
 };
 
 Factory.prototype = {
     constructor: Factory,
 
     /**
+     * Returns constructed object.
      *
-     * @param definition
+     * @abstract
+     *
+     * @param {Object} definition
      *
      * @returns {Object|Promise}
      */
     factory: function(definition) {
-        throw new Error("Method must be implemented");
+        throw new Error("Method 'factory' must be implemented");
     }
 };
 
@@ -23,5 +32,6 @@ Factory.extend = function(prots, statics) {
     return inherits(this, prots, statics);
 };
 
+Factory.DEFAULTS = {};
 
 module.exports = Factory;