+++ src/dm/parser/string/template/parameter-live.js
@@ -17,7 +17,7 @@ ParameterLiveTemplate = Template.extend(
         define: function(match) {
             return {
                 name: match[1]
-            }
+            };
         }
     },
     {