+++ src/dm/parser/string/template/parameter.js
@@ -17,7 +17,7 @@ ParameterTemplate = Template.extend(
         define: function(match) {
             return {
                 name: match[1]
-            }
+            };
         }
     },
     {